requirements.txt
pwntools @ git+https://github.com/arthaud/python3-pwntools.git@61210a68cd88e9084c72292d3119c38c44f07966
colorama==0.4.5
requests==2.27.1